webhooks: move GraphQL resolver to enterprise package.
Test plan:
Existing test cases should pass after refactoring.
