+++ enterprise/cmd/frontend/internal/repos/init.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/enterprise"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/repos/webhooks"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/repos/webhooks/resolvers"
 	"github.com/sourcegraph/sourcegraph/internal/codeintel"
 	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
 	"github.com/sourcegraph/sourcegraph/internal/database"
@@ -13,7 +14,7 @@ import (
 
 // Init initializes the given enterpriseServices with the webhook handlers for handling GitHub push events.
 func Init(
-	ctx context.Context,
+	_ context.Context,
 	db database.DB,
 	_ codeintel.Services,
 	_ conftypes.UnifiedWatchable,
@@ -21,5 +22,6 @@ func Init(
 	_ *observation.Context,
 ) error {
 	enterpriseServices.GitHubSyncWebhook = webhooks.NewGitHubWebhookHandler()
+	enterpriseServices.WebhooksResolver = resolvers.NewWebhooksResolver(db)
 	return nil
 }
